mode_choice macro dummies replaced with annotations
