+++ activitysim/abm/models/__init__.py
@@ -17,12 +17,14 @@ import non_mandatory_destination
 import non_mandatory_scheduling
 import school_location
 import workplace_location
-import mode_choice
+import tour_mode_choice
 import cdap
 import create_trips
+import trip_mode_choice
 import atwork_subtour_frequency
 import atwork_subtour_destination
 import atwork_subtour_scheduling
+import atwork_subtour_mode_choice
 
 import utility_steps
