+++ activitysim/abm/models/non_mandatory_scheduling.py
@@ -13,8 +13,9 @@ from activitysim.core import inject
 from activitysim.core import pipeline
 from activitysim.core import timetable as tt
 
-
+from .util import expressions
 from .util.vectorize_tour_scheduling import vectorize_tour_scheduling
+from activitysim.core.util import assign_in_place
 
 
 logger = logging.getLogger(__name__)
@@ -61,12 +62,10 @@ def non_mandatory_tour_scheduling(tours,
         chunk_size=chunk_size,
         trace_label=trace_label)
 
-    # add tdd_choices columns to tours
-    for c in tdd_choices.columns:
-        tours.loc[tdd_choices.index, c] = tdd_choices[c]
-
+    assign_in_place(tours, tdd_choices)
     pipeline.replace_table("tours", tours)
 
+    # updated df for tracing
     non_mandatory_tours = tours[tours.non_mandatory]
 
     tracing.dump_df(DUMP,
