+++ example/configs/annotate_tours_mode_choice_preprocessor1.csv
@@ -1,0 +1,15 @@
+Description,Target,Expression
+#,,
+#,, auto peak
+#,,
+local,_SKIM_TIME_PERIODS,"setting('skim_time_periods')"
+local,_SKIM_PERIODS,"_SKIM_TIME_PERIODS['hours']"
+local,_SKIM_LABELS,"_SKIM_TIME_PERIODS['labels']"
+local,_IS_JOINT,"'number_of_participants' in df.columns"
+local,_HAVE_PARENT,"'parent_tour_id' in df.columns"
+out_period skim label(as string not categorical),out_period,"pd.cut(df.end, _SKIM_PERIODS, labels=_SKIM_LABELS).astype(str)"
+in_period skim label(as string not categorical),in_period,"pd.cut(df.start, _SKIM_PERIODS, labels=_SKIM_LABELS).astype(str)"
+dest_density_index,dest_density_index,"reindex(land_use.density_index, df.destination)"
+dest_topology,dest_topology,"reindex(land_use.TOPOLOGY, df.destination)"
+#,, attention! mode is a dataframe method so we cant say tours.mode
+,parent_tour_mode,"reindex(tours['mode'], df.parent_tour_id) if _HAVE_PARENT else ''"
