+++ example/configs/annotate_tours_mode_choice_preprocessor2.csv
@@ -1,0 +1,32 @@
+Description,Target,Expression
+#,,
+#,, auto peak
+#,,
+local,_IS_JOINT,"'number_of_participants' in tours.columns"
+,number_of_participants,"tours.number_of_participants if _IS_JOINT else 1"
+,is_joint,(number_of_participants > 1)
+#FIXME note that there's something about whether this is a subtour?,work_tour_is_drive,0
+,work_tour_is_drive,"tours.parent_tour_mode.isin(['DRIVEALONEFREE','DRIVEALONEPAY'])"
+,work_tour_is_bike,"tours.parent_tour_mode=='BIKE'"
+,terminal_time,"reindex(land_use.TERMINAL, tours[dest_col_name])"
+FIXME,origin_walk_time,0
+FIXME,destination_walk_time,0
+FIXME,daily_parking_cost,0
+#,,
+,drive_commuter_available,(odt_skims['DRV_COM_WLK_TOTIVT']>0) & (dot_skims['WLK_COM_DRV_TOTIVT']>0) & ((odt_skims['DRV_COM_WLK_KEYIVT']>0) | (dot_skims['WLK_COM_DRV_KEYIVT']>0))
+,drive_express_available,(odt_skims['DRV_EXP_WLK_TOTIVT']>0) & (dot_skims['WLK_EXP_DRV_TOTIVT']>0) & ((odt_skims['DRV_EXP_WLK_KEYIVT']>0) | (dot_skims['WLK_EXP_DRV_KEYIVT']>0))
+,drive_heavyrail_available,(odt_skims['DRV_HVY_WLK_TOTIVT']>0) & (dot_skims['WLK_HVY_DRV_TOTIVT']>0) & ((odt_skims['DRV_HVY_WLK_KEYIVT']>0) | (dot_skims['WLK_HVY_DRV_KEYIVT']>0))
+,drive_local_available,(odt_skims['DRV_LOC_WLK_TOTIVT']>0) & (dot_skims['WLK_LOC_DRV_TOTIVT']>0)
+,drive_lrf_available,(odt_skims['DRV_LRF_WLK_TOTIVT']>0) & (dot_skims['WLK_LRF_DRV_TOTIVT']>0) & ((odt_skims['DRV_LRF_WLK_KEYIVT']>0) | (dot_skims['WLK_LRF_DRV_KEYIVT']>0))
+,hov2_available,(odt_skims['HOV2_TIME']>0) & (dot_skims['HOV2_TIME']>0)
+,hov2toll_available,(odt_skims['HOV2TOLL_VTOLL']>0) | (dot_skims['HOV2TOLL_VTOLL']>0)
+,hov3_available,(odt_skims['HOV3_TIME']>0) & (dot_skims['HOV3_TIME']>0)
+,hov3toll_available,(odt_skims['HOV3TOLL_VTOLL']>0) | (dot_skims['HOV3TOLL_VTOLL']>0)
+,sov_available,(odt_skims['SOV_TIME']>0) & (dot_skims['SOV_TIME']>0)
+,sovtoll_available,(odt_skims['SOVTOLL_VTOLL']>0) | (dot_skims['SOVTOLL_VTOLL']>0)
+,walk_commuter_available,(odt_skims['WLK_COM_WLK_TOTIVT']>0) & (dot_skims['WLK_COM_WLK_TOTIVT']>0) & ((odt_skims['WLK_COM_WLK_KEYIVT']>0) | (dot_skims['WLK_COM_WLK_KEYIVT']>0))
+,walk_express_available,(odt_skims['WLK_EXP_WLK_TOTIVT']>0) & (dot_skims['WLK_EXP_WLK_TOTIVT']>0) & ((odt_skims['WLK_EXP_WLK_KEYIVT']>0) | (dot_skims['WLK_EXP_WLK_KEYIVT']>0))
+,walk_heavyrail_available,(odt_skims['WLK_HVY_WLK_TOTIVT']>0) & (dot_skims['WLK_HVY_WLK_TOTIVT']>0) & ((odt_skims['WLK_HVY_WLK_KEYIVT']>0) | (dot_skims['WLK_HVY_WLK_KEYIVT']>0))
+,walk_local_available,(odt_skims['WLK_LOC_WLK_TOTIVT']>0) & (dot_skims['WLK_LOC_WLK_TOTIVT']>0)
+,walk_lrf_available,(odt_skims['WLK_LRF_WLK_TOTIVT']>0) & (dot_skims['WLK_LRF_WLK_TOTIVT']>0) & ((odt_skims['WLK_LRF_WLK_KEYIVT']>0) | (dot_skims['WLK_LRF_WLK_KEYIVT']>0))
+,distance,od_skims['DIST']
