+++ example/configs/tour_mode_choice.yaml
@@ -67,3 +67,17 @@ CONSTANTS:
     upperAM: 10
     upperMD: 15
     upperPM: 19
+
+
+# so far, we can use the same spec as for non-joint tours
+preprocessor_settings:
+  - SPEC: annotate_tours_mode_choice_preprocessor1
+    DF: df
+    TABLES:
+      - land_use
+      - tours
+
+  - SPEC: annotate_tours_mode_choice_preprocessor2
+    DF: tours
+    TABLES:
+      - land_use
